Extract reviews API into separate route file and fix rating null response
